core code update

+++ src/block.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Block {
-    pub index: u64,
+    pub index: usize,
     pub timestamp: DateTime<Utc>,
     pub choice: String,
     pub voter: String,
